+++ docs/changelog.rst
@@ -30,6 +30,8 @@ In progress (unreleased)
   :class:`pickle.Pickler`
 * :gh:issue:`1430` :mod:`mitogen`: Remove caching of result of
   :meth:`mitogen.core.Message.unpickle`
+* :gh:issue:`1430` :mod:`mitogen`: Allow mutiple pickle streams in a single
+  :class:`mitogen.core.Message`, add :meth:`mitogen.core.Message.unpickle_iter`
 
 
 v0.3.40 (2026-02-04)