+++ docs/changelog.rst
@@ -32,6 +32,8 @@ In progress (unreleased)
   :meth:`mitogen.core.Message.unpickle`
 * :gh:issue:`1430` :mod:`mitogen`: Allow mutiple pickle streams in a single
   :class:`mitogen.core.Message`, add :meth:`mitogen.core.Message.unpickle_iter`
+* :gh:issue:`1430` :mod:`mitogen`: Speed up :class:`mitogen.core.ResourceReader`
+  using 2 pickle streams in :data:`mitogen.core.LOAD_RESOURCE` messages
 
 
 v0.3.40 (2026-02-04)